Remove Sprite.Scale in favor of a new SetScale+GetScale
The single Scale field was convenient, but insufficient.
Sometimes we need to scale one dimension but not another.

Using a method instead of a fields here to make it easier
to avoid breaking the API in the future.

The games that use `s.Scale = x` should be changed to `s.SetScale(x, x)`.